Updates README.md with command-line examples
This is an intermediate update of the `README.md`,
that means, that it is not finished yet; but there are yet so many
changes that it already might be beneficial to have it available for all
users (even if's not finished).

We're aiming to provice extensive examples and descriptions for all the
commands that are used in this Workshop/Tutorial.